AgentServer/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
openai>=1.35.3
mcp>=1.2.0
python-dotenv>=1.0.1
